1.12.2-3.1.8
Added textures from the defunct Extra Buttons mod and created sticky buttons that act in the same way. When lit they push a level 15 redstone signal, when off they don't. Must be pressed to toggle.

Still need to adjust the model hit boxes.

If anyone knows how to make it better feel free to fork it and submit a pull request.